fix(KG): SIGReg unit-vector bug + ExecutionStep field docs/schema nav
- optimization_engine._sigreg_normality_test: drop erroneous float() that collapsed the
  d-dim random projection vector to a scalar ('only 0-dimensional arrays can be converted')
  — broke all 4 latent-space collapse-detection tests.
- models/sdd.Task: give refined_subtask (CONCEPT:ORCH-1.1) and access_list (CONCEPT:ORCH-1.3)
  proper Field(description=...) — they were description=None, so the LLM-facing schema and
  the traceability tests had nothing to assert.
- test_visibility_graph / test_conductor_workflow: ExecutionStep(==Task) is self-referential
  (subtasks), so pydantic emits properties under $defs behind a top-level $ref; resolve the
  ref instead of indexing schema['properties'] (KeyError).

+++ agent_utilities/knowledge_graph/memory/optimization_engine.py
@@ -356,7 +356,10 @@ class MemoryOptimizationEngine:
         min_p = 1.0
         for _ in range(self._n_projections):
             direction = rng.standard_normal(d)
-            direction = float(direction / (np.linalg.norm(direction) or 1.0))
+            # Normalise the random projection vector to unit length; it must stay a
+            # d-dim array (wrapping it in float() collapsed it and raised
+            # "only 0-dimensional arrays can be converted to Python scalars").
+            direction = direction / (np.linalg.norm(direction) or 1.0)
             projection = centered @ direction
             p_value = self._simplified_normality_p(projection)
             min_p = float(min(min_p, p_value))
